requirements.txt
# requirements.txt
requests == 2.31.0
polars == 0.20.15
duckdb == 0.9.2
duckdb-engine == 0.9.2
python.dotenv == 1.0.0
schedule == 1.2.1
flask == 3.1.0
flask_cors == 5.0.1
pyarrow == 19.0.1